% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggortho.R
\name{construct_ggortho}
\alias{construct_ggortho}
\alias{ggortho}
\alias{ggortho2}
\title{Plot Orthographic view of an image using \code{ggplot2}}
\usage{
construct_ggortho(img, xyz = NULL, ...)

ggortho(img, xyz = NULL, ...)

ggortho2(...)
}
\arguments{
\item{img}{object that can be coerced to a \code{nifti} using
\code{\link{check_nifti}}}

\item{xyz}{coordinates to be plotted}

\item{...}{arguments to pass to \code{\link{img_colour_df}}}
}
\value{
\code{construct_ggortho} returns a list of the constructed plot,
the full data for the image, the image for the plot and the \code{xyz} coordinate.
\code{ggortho} simply returns the plot
}
\description{
Plotting function for an orthographic view of a 3-dimensional image
}
\note{
\code{ggortho2} is a duplicate of \code{ggortho} to agree with the
\code{\link{ortho2}} of \code{neurobae}
}
\examples{
img = oro.nifti::nifti(array(rnorm(10^3), dim = rep(10, 3)))
L = construct_ggortho(img)
ggortho(img)
}
