% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ortho.R
\name{theme_ortho}
\alias{theme_ortho}
\title{Theme for orthographic plots in \code{ggplot2}}
\usage{
theme_ortho()
}
\value{
A \code{ggplot2} theme
}
\description{
This is a set up of a theme for \code{ggplot2}, specifically for
orthographic plots
}
\examples{
theme_ortho()
}
